v1.1.0 - Improve auto-downloading & LWJGL 3
Add new config option, timeout, and remove redundant downloads for LWJGL 3. As it says on their documentation, there is no need to specify .so files if the correct natives are used. So there is also no need to specify the so file paths for LWJGL 3. Still vital to fixing LWJGL 2 though.